chore: vitestBaseConfig 추가
