+++ packages/__pub-testing__/vitest.config.mts
@@ -1,17 +1,3 @@
-import { resolve } from 'node:path';
-import { findUpSync } from 'find-up';
+import { vitestBaseConfig } from '../../scripts/vitestBaseConfig';
 
-import { defineConfig } from 'vitest/config';
-
-const rootDir = (findUpSync('nx.json') as string).replace(
-  /(.*)([\/\\])nx.json$/,
-  '$1'
-);
-
-export default defineConfig({
-  test: {
-    globals: true,
-    environment: 'jsdom',
-    setupFiles: resolve(rootDir, './scripts/setup-tests.ts'),
-  },
-});
+export default vitestBaseConfig;
